Tests/LibWeb/Text/expected/Compression/DecompressionStream.txt

format=deflate-raw: Well hello friends!
format=gzip: Well hello friends!
